refactor packages inside redux module
