+++ src/main/java/redux/store/Notifier.java
@@ -2,6 +2,7 @@ package redux.store;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+import redux.state.State;
 
 public interface Notifier<S extends State<A>, A> {
     Mono<S> getLatestState$();
